+++ client/admin-plugin-client-plugin.ts
@@ -211,11 +211,18 @@ function register ({ registerHook, registerSettingsScript, peertubeHelpers }: Re
         case 'prosody-muc-log-by-default':
         case 'prosody-muc-expiration':
         case 'prosody-c2s':
+        case 'prosody-components':
           return options.formValues['chat-type'] !== ('builtin-prosody' as ChatType)
         case 'prosody-c2s-port':
           return !(
             options.formValues['chat-type'] === ('builtin-prosody' as ChatType) &&
             options.formValues['prosody-c2s'] === true
+          )
+        case 'prosody-components-port':
+        case 'prosody-components-list':
+          return !(
+            options.formValues['chat-type'] === ('builtin-prosody' as ChatType) &&
+            options.formValues['prosody-components'] === true
           )
         case 'chat-server':
         case 'chat-room':